Solution for multi server

+++ reprocess_game_history.py
@@ -4,10 +4,12 @@ from sqlalchemy.orm import with_expression
 from rcon.models import enter_session
 from rcon.workers import _record_stats
 from rcon.utils import MapsHistory
+import sys
 
 with enter_session() as sess:
-    sess.execute('truncate table map_history cascade')
-    sess.commit()
+    if not sys.argv[-1] == 'skiperase':
+        sess.execute('truncate table map_history cascade')
+        sess.commit()
     for m in MapsHistory():
         try:
             if m['start'] and m['end']:
